code reorganization, much clearer now
