+++ .gitignore
@@ -1,12 +1,8 @@
-
 *.js
 !gulpfile.js
 *.js.map
 
 typings/
-
-
-
 
 
 ###### added by default by Git
