+++ gulpfile.js
@@ -27,7 +27,7 @@ var source = require('vinyl-source-stream');
 function bundleShare(b, name) {
     return new Promise(function(resolve, reject){
         b.bundle()
-        .pipe(source( join('.', 'client', name) ) )
+        .pipe(source( join('.', name)))
         .pipe(gulp.dest('.'))
         .on('error', function (err) {
             console.error('bundleShare error', err.message);
@@ -48,135 +48,69 @@ function browserifyShare(name){
         debug: true
     });
     
-    b.add( join('.', name, 'src', 'main.js') );
+    b.add( join('.', 'js', name, 'main.js') );
 
     return bundleShare(b, 'browserify-bundle.js');
 }
 
-function tscDev(name){
+function tscDev(){
 
-    return gulp.src(join(name, '**/*.ts'))
+    return gulp.src(join('ts', '**/*.ts'))
         .pipe(sourcemaps.init())
         .pipe(ts({
             noImplicitAny: true,
             target: 'ES5',
-            module: 'commonjs'
+            module: 'commonjs',
         }))
         .pipe(sourcemaps.write())
-        .pipe(gulp.dest(join('.', name)));
+        .pipe(gulp.dest(join('.', 'js')));
 }
 
-function tscProd(name){ // doesn't generate sourcemaps
+function tscProd(){ // doesn't generate sourcemaps
 
-    return gulp.src(join(name, '**/*.ts'))
+    return gulp.src(join('ts', '**/*.ts'))
         .pipe(ts({
             noImplicitAny: true,
             target: 'ES5',
             module: 'commonjs'
         }))
-        .pipe(gulp.dest(join('.', name)));
+        .pipe(gulp.dest(join('.', 'js')));
 }
 
-// Browserify bundle the client
-gulp.task('browserify-client', function(){
-    return browserifyShare('client');
-});
-
 // Typescript compile and then browserify bundle the client
-gulp.task('build-client-dev', ['tsc-client-dev'], function(){
+gulp.task('build-dev', ['tsc-dev'], function(){
     return browserifyShare('client');
 });
 
-gulp.task('build-client-prod', ['tsc-client-prod'], function(){
+gulp.task('build-prod', ['tsc-prod'], function(){
     return browserifyShare('client');
 });
 
-// Typescript compile the client
-gulp.task('tsc-client-dev', function(){
-    return tscDev('client');
+// Typescript compile
+gulp.task('tsc-dev', function(){
+    return tscDev();
 });
 
-gulp.task('tsc-client-prod', function(){
-    return tscProd('client');
-});
-
-// Typescript compile the server
-gulp.task('tsc-server-dev', function(){
-    return tscDev('server');
-});
-
-gulp.task('tsc-server-prod', function(){
-    return tscProd('server');
-});
-
-// Typescript compile the tools
-gulp.task('tsc-tools-dev', function(){
-    return tscDev('tools');
-});
-
-gulp.task('tsc-tools-prod', function(){
-    return tscProd('tools');
-});
-
-// Typescript compile the tests
-gulp.task('tsc-tests-dev', function(){
-    return tscDev('tests');
+gulp.task('tsc-prod', function(){
+    return tscProd();
 });
 
 // Watch client
-gulp.task('watch-client', ['build-dev'], function() {
+gulp.task('watch', ['build-dev'], function() {
     console.log('Watching client');
     
     // When a .js updates, Browserify bundle the client
-    var jsClientWatcher = gulp.watch('./client/src/**/*.js', ['browserify-client']);
+    var jsClientWatcher = gulp.watch('./js/client/**/*.js', ['browserify-client']);
     jsClientWatcher.on('change', function(event) {
         console.log('** JS Client ** File ' + path.relative(__dirname, event.path) + ' was ' + event.type);
     });
 
     // When a .ts updates, Typescript compile the client
-    var tsClientWatcher = gulp.watch('./client/src/**/*.ts', ['tsc-client-dev']);
+    var tsClientWatcher = gulp.watch('./ts/**/*.ts', ['tsc-client-dev']);
     tsClientWatcher.on('change', function(event) {
         console.log('** TS Client ** File ' + path.relative(__dirname, event.path) + ' was ' + event.type);
     });
 });
-
-// Watch server
-gulp.task('watch-server', ['build-dev'], function() {
-    console.log('Watching server');
-
-    // When a .ts updates, Typescript compile the server
-    var tsServerWatcher = gulp.watch('./server/src/**/*.ts', ['tsc-server-dev']);
-    tsServerWatcher.on('change', function(event) {
-        console.log('** TS Server ** File ' + path.relative(__dirname, event.path) + ' was ' + event.type);
-    });
-});
-
-// Watch tools
-gulp.task('watch-tools', ['build-dev'], function() {
-    console.log('Watching tools');
-
-    // When a .ts updates, Typescript compile the server
-    var tsToolsWatcher = gulp.watch('./tools/**/*.ts', ['tsc-tools-dev']);
-    tsToolsWatcher.on('change', function(event) {
-        console.log('** TS Tools ** File ' + path.relative(__dirname, event.path) + ' was ' + event.type);
-    });
-});
-
-// Watch tests
-gulp.task('watch-tests', ['build-dev'], function() {
-    console.log('Watching tests');
-
-    // When a .ts updates, Typescript compile the server
-    var tsTestsWatcher = gulp.watch('./tests/**/*.ts', ['tsc-tests-dev']);
-    tsTestsWatcher.on('change', function(event) {
-        console.log('** TS Tests ** File ' + path.relative(__dirname, event.path) + ' was ' + event.type);
-    });
-});
-
-gulp.task('watch', ['watch-client', 'watch-server', 'watch-tools', 'watch-tests']);
-gulp.task('build-dev', ['build-client-dev', 'tsc-server-dev', 'tsc-tools-dev', 'tsc-tests-dev']);
-gulp.task('build-prod', ['build-client-prod', 'tsc-server-prod', 'tsc-tools-prod']);
-
 
 // These are to spawn docker test instances, needed mainly for async tests that require the server being up
 
@@ -188,8 +122,6 @@ gulp.task('start-containers-test', function(){
     Top-level tasks
 */
 
-gulp.task('dev', ['start-containers-dev', 'watch']);
-gulp.task('prod', ['start-containers-prod']);
 gulp.task('test', ['start-containers-test']);
