+++ ts/client/Components/Smart/Application.ts
