+++ ts/tools/tryConnectTo.ts
@@ -1,4 +1,4 @@
-/// <reference path="../typings/tsd.d.ts" />
+/// <reference path="../../typings/tsd.d.ts" />
 'use strict';
 
 import * as request from 'request';
